moving client into zpy
